+++ packages/elvis/CHANGELOG.md
@@ -10,7 +10,7 @@
       </span>
       <br>
       <span>
-        December 10, 2021
+        December 14, 2021
       </span>
     </td>
     <td class="changelog-content">
@@ -18,10 +18,14 @@
         <div class="changelog-section-title">🐝 Bug Fixes</div>
           <div>Component: 
             <a class="e-link e-link--inline" href="https://design.elvia.io/components/card#Overview">Card</a> &
+            <a class="e-link e-link--inline" href="https://design.elvia.io/components/link#Overview">Link</a>
           </div>
           <ul class="e-list">
             <li>
-              Deprecated e-card, replaced with card component.
+              Deprecated e-card and e-link--card, replaced with card component.  
+            </li>
+            <li>
+              Deprecated e-link--jumbo, replaced with  <a class="e-link e-link--inline" href="https://design.elvia.io/patterns/groups#Examples">action group pattern</a>.  
             </li>
           </ul>
         </div>